created twitter helpers file which holds functions and is imported
